Get tests working again
