+++ tests/fixtures/__init__.py
@@ -1,0 +1,2 @@
+from .image import *
+from .misc import *
